Refine about page chrome and background styling
